requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
langchain-google-genai==0.0.11
langchain-core==0.1.16

# Vector database
chromadb==0.4.22

# LLM Providers
openai==1.10.0
google-generativeai==0.4.1
ollama>=0.1.0

# HuggingFace (optional, for local embeddings)
sentence-transformers==2.3.1

# Document loaders
pypdf==4.0.1
python-docx==1.1.0
unstructured==0.12.0
markdown==3.5.2

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.3

# CLI and display
colorama==0.4.6

# Testing and Evaluation
langchain-mistralai>=0.1.0
ragas>=0.2.0
datasets>=2.14.0
